% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saeall_sample.R
\docType{data}
\name{saeall_sample}
\alias{saeall_sample}
\title{saeall sample data}
\format{
A list containing:
\describe{
\item{data}{DataFrame of the original data}
\item{Gamma}{Gamma matrix with partition A and B}
\item{sigma_e}{Sigma e variance}
\item{sigma_u}{Sigma u variance}
}
}
\usage{
data(saeall_sample)
}
\description{
Data is generated using MIND package (D’Alò et al., 2021) from a sample of 10,000 individuals a
nd 104 municipalities selected from the second one, including a population of 514,320 individuals and 333 towns.
We considered a sub-set of the sample data referred to
49 municipalities to reduce computational time.
Each sample record has the following variables:
-Sex by age class (Sexage, 14 categories).
-Educational level (Edu, six levels),
-Nationality (Fore, 1=italian, 2=otherwise),
-Municipality code (Mun, 49 codes),
-Employment status (Emp, 1=employed and 0=otherwise).
We randomly selected 24 municipalities,
and we assigned to one of two Macro-strata, say Macro-strata 1,
while we set the remaining 25 towns to the other Macro-stratum, say Macro-stratum2.
After this process, the dataset has two types of overlapping domains.
The first type, denoted as partition A, has 49 domains corresponding to a municipality.
In contrast, the second partition, marked as partition B, has 2 domains,
each of which is a macro-stratum. The domains of the B partition are an aggregation of type A partition domains
}
\keyword{datasets}
